requirements.txt
bios
torchvision
fc_pruning==0.0.16
